+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Patches / Fixes
+
+* Fix - #5780 - Fix inconsistent image series form behavior when changing patient or visit, by removing the patient and visit fields
+
 ## 7.2.1
 
 ### Patches / Fixes